+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- Added support for 'method' parameter on Replacer 'add_rule'
+
 ## [0.5.0] - 2025-12-15
 ### Added
 - Add the APIs of the following add-ons: